ddl: split series test into new package
